Fix timeSpeed bug
Basically, when multiplying the velocity each time by also time speed,
the position started, after an arbitrary amount of iterations, to get
unprecise, as empirically calculating the Force applied by gravity
only works when small steps in time are done (it's the same line of
reasoning as for integrals).
If time speed values were too great, then the pendulum started to move
away from the string.

To solve the problem I decided not to multiply deltaTime by timeSpeed,
but to instead only scale the velocity by the accelleration*deltaTime
(aka only doing small steps in time). Furthermore, I repeat this
step timeSpeed times in each frame, so that, even though more
computations are required, the amount of precision doesn't rely on
timeSpeed.